refactor(settings): migrate cloud_print into a Settings Section (closes the controller bypass)
delete_all_settings() now covers cloud_print (deliberate fix); grep confirms zero callers in free and Pro. get_settings('cloud_print') now returns the redacted view instead of WP_Error (improvement; zero callers relied on the error).

Notes:
- sanitize_cloud_assignment() kept on controller (private, phpstan-suppressed) because Settings_CloudPrint_Test conformance gate uses ReflectionMethod to call it directly; production path fully delegates to Cloud_Print_Section.
- Removed sanitize_cloud_printer and with_cloud_printer_encoding_fields from controller (no external test references).
- Cloud_Print_Registry and Provider imports removed from API/Settings.php (zero remaining references).
