+++ includes/API/Settings.php
@@ -8,8 +8,6 @@
 namespace WCPOS\WooCommercePOS\API;
 
 use Closure;
-use WCPOS\WooCommercePOS\Services\Cloud_Print_Registry;
-use WCPOS\WooCommercePOS\Services\Provider;
 use WCPOS\WooCommercePOS\Services\Settings as SettingsService;
 use WCPOS\WooCommercePOS\Services\Tax_Id_Detector;
 use WCPOS\WooCommercePOS\Services\Tax_Id_Settings;
@@ -501,33 +499,10 @@ class Settings extends WP_REST_Controller {
 	 * @return WP_REST_Response
 	 */
 	public function get_cloud_print_settings() {
-		$settings = get_option( 'woocommerce_pos_settings_cloud_print', array() );
-		$settings = wp_parse_args(
-			\is_array( $settings ) ? $settings : array(),
-			array(
-				'printers'    => array(),
-				'assignments' => array(),
-			)
+		return new WP_REST_Response(
+			SettingsService::instance()->sections()->get( 'cloud_print' )->read(),
+			200
 		);
-		$registry             = new Cloud_Print_Registry();
-		$settings['printers'] = array_map(
-			function ( $printer ) use ( $registry ) {
-				if ( ! \is_array( $printer ) ) {
-					return $printer;
-				}
-				$id                   = (string) ( $printer['id'] ?? '' );
-				$seen                 = $registry->get_seen( $id );
-				$printer              = $this->with_cloud_printer_encoding_fields( $printer );
-				$printer['status']    = $registry->status_for( $id );
-				$printer['last_seen'] = $seen > 0 ? $seen : null;
-				unset( $printer['poll_token_hash'], $printer['printnode_api_key'], $printer['star_api_key'] );
-
-				return $printer;
-			},
-			$settings['printers']
-		);
-
-		return new WP_REST_Response( $settings, 200 );
 	}
 
 	/**
@@ -542,214 +517,46 @@ class Settings extends WP_REST_Controller {
 		if ( empty( $payload ) ) {
 			$payload = $request->get_body_params();
 		}
-		$printers = isset( $payload['printers'] ) && \is_array( $payload['printers'] ) ? array_values( $payload['printers'] ) : array();
-		$assigns  = isset( $payload['assignments'] ) && \is_array( $payload['assignments'] ) ? array_values( $payload['assignments'] ) : array();
 
-		$existing        = get_option( 'woocommerce_pos_settings_cloud_print', array() );
-		$existing_hashes = array();
-		$existing_keys      = array();
-		$existing_star_keys = array();
-		$existing_ids       = array();
-		if ( isset( $existing['printers'] ) && \is_array( $existing['printers'] ) ) {
-			foreach ( $existing['printers'] as $printer ) {
-				if ( ! empty( $printer['id'] ) ) {
-					$existing_ids[] = $printer['id'];
-				}
-				if ( ! empty( $printer['id'] ) && ! empty( $printer['poll_token_hash'] ) ) {
-					$existing_hashes[ $printer['id'] ] = $printer['poll_token_hash'];
-				}
-				if ( ! empty( $printer['id'] ) && ! empty( $printer['printnode_api_key'] ) ) {
-					$existing_keys[ $printer['id'] ] = $printer['printnode_api_key'];
-				}
-				if ( ! empty( $printer['id'] ) && ! empty( $printer['star_api_key'] ) ) {
-					$existing_star_keys[ $printer['id'] ] = $printer['star_api_key'];
-				}
-			}
-		}
+		$result = SettingsService::instance()->sections()->get( 'cloud_print' )->write( (array) $payload );
 
-		$generated      = array();
-		$clean_printers = array();
-		$seen_ids       = array();
-		foreach ( $printers as $printer ) {
-			$printer = $this->sanitize_cloud_printer( $printer );
-
-			if ( '' === $printer['id'] ) {
-				$printer['id'] = Cloud_Print_Registry::derive_id( $printer['name'], array_merge( $existing_ids, array_keys( $seen_ids ) ) );
-			}
-			$id = $printer['id'];
-
-			// Preserve a previously stored PrintNode API key when the incoming
-			// payload omits it (GET strips the key, so the React app re-POSTs
-			// printers without it when toggling other fields). A non-empty
-			// incoming key still overwrites, letting users rotate it.
-			if ( 'printnode' === $printer['provider'] && '' === $printer['printnode_api_key'] && ! empty( $existing_keys[ $id ] ) ) {
-				$printer['printnode_api_key'] = $existing_keys[ $id ];
-			}
-
-			if ( 'star-online' === $printer['provider'] && '' === $printer['star_api_key'] && ! empty( $existing_star_keys[ $id ] ) ) {
-				$printer['star_api_key'] = $existing_star_keys[ $id ];
-			}
-
-			if ( 'star-online' === $printer['provider'] ) {
-				$api_base = \WCPOS\WooCommercePOS\Services\Star_Online_Client::api_base_from_cloudprnt_url( (string) $printer['star_cloudprnt_url'] );
-				$group    = \WCPOS\WooCommercePOS\Services\Star_Online_Client::group_from_cloudprnt_url( (string) $printer['star_cloudprnt_url'] );
-				if ( '' === $printer['star_api_key'] || null === $api_base || '' === $group || '' === $printer['star_device_id'] ) {
-					return new WP_REST_Response(
-						array(
-							'code'    => 'wcpos_cloud_print_star_online_invalid',
-							'message' => __( 'Star Online printers need an API key, a valid stario.online CloudPRNT URL, and a device.', 'woocommerce-pos' ),
-						),
-						400
-					);
-				}
-			}
-
-			if ( isset( $seen_ids[ $id ] ) ) {
-				return new WP_REST_Response(
-					array(
-						'code'    => 'wcpos_cloud_print_duplicate_printer_id',
-						'message' => __( 'Duplicate printer id.', 'woocommerce-pos' ),
-					),
-					400
-				);
-			}
-			$seen_ids[ $id ] = true;
-
-			$regenerate = ! empty( $printer['regenerate_token'] );
-			unset( $printer['regenerate_token'] );
-
-			if ( Provider::is_polling( $printer['provider'] ) ) {
-				if ( $regenerate || empty( $existing_hashes[ $id ] ) ) {
-					$token                      = Cloud_Print_Registry::generate_token();
-					$printer['poll_token_hash'] = Cloud_Print_Registry::hash_token( $token );
-					$generated[ $id ]           = $token;
-				} else {
-					$printer['poll_token_hash'] = $existing_hashes[ $id ];
-				}
-			}
-
-			$clean_printers[] = $printer;
-		}
-
-		$clean = array(
-			'printers'    => $clean_printers,
-			'assignments' => array_map( array( $this, 'sanitize_cloud_assignment' ), $assigns ),
-		);
-		update_option( 'woocommerce_pos_settings_cloud_print', $clean );
-
-		// Drop runtime last-seen entries for printers that were removed.
-		( new Cloud_Print_Registry() )->prune_seen( array_keys( $seen_ids ) );
-
-		$response_printers = array_map(
-			function ( $printer ) {
-				$printer = $this->with_cloud_printer_encoding_fields( $printer );
-				unset( $printer['poll_token_hash'], $printer['printnode_api_key'], $printer['star_api_key'] );
-
-				return $printer;
-			},
-			$clean_printers
-		);
-
-		return new WP_REST_Response(
-			array(
-				'printers'    => $response_printers,
-				'assignments' => $clean['assignments'],
-				'generated'   => $generated,
-			),
-			200
-		);
-	}
-
-	/**
-	 * Sanitize a cloud printer entry.
-	 *
-	 * @param mixed $printer Printer.
-	 *
-	 * @return array
-	 */
-	private function sanitize_cloud_printer( $printer ): array {
-		$printer  = \is_array( $printer ) ? $printer : array();
-		$provider = \in_array( $printer['provider'] ?? '', Provider::valid(), true )
-			? $printer['provider'] : 'star-cloudprnt';
-
-		$clean = array(
-			'id'               => sanitize_text_field( $printer['id'] ?? '' ),
-			'name'             => sanitize_text_field( $printer['name'] ?? '' ),
-			'provider'         => $provider,
-			'store_id'         => isset( $printer['store_id'] ) ? (int) $printer['store_id'] : 0,
-			'regenerate_token' => ! empty( $printer['regenerate_token'] ),
-		);
-		if ( 'printnode' === $provider ) {
-			$clean['printnode_api_key']    = sanitize_text_field( $printer['printnode_api_key'] ?? '' );
-			$clean['printnode_printer_id'] = isset( $printer['printnode_printer_id'] ) ? (int) $printer['printnode_printer_id'] : 0;
-			$clean['printnode_format']     = \in_array( $printer['printnode_format'] ?? '', array( 'pdf', 'raw' ), true )
-				? $printer['printnode_format'] : 'pdf';
-		}
-		if ( 'star-cloudprnt' === $provider ) {
-			$encoding_fields = array_intersect_key(
-				$printer,
-				array_flip( array( 'columns', 'language', 'autoCut', 'fullReceiptRaster' ) )
-			);
-			$clean           = $this->with_cloud_printer_encoding_fields(
-				array_merge( $clean, $encoding_fields )
+		if ( is_wp_error( $result ) ) {
+			// Keep the historical error body shape {code, message} — clients do
+			// not expect WP_Error's extra data envelope here.
+			return new WP_REST_Response(
+				array(
+					'code'    => $result->get_error_code(),
+					'message' => $result->get_error_message(),
+				),
+				400
 			);
 		}
-		if ( 'star-online' === $provider ) {
-			$clean['star_api_key']       = sanitize_text_field( $printer['star_api_key'] ?? '' );
-			$clean['star_cloudprnt_url'] = esc_url_raw( $printer['star_cloudprnt_url'] ?? '' );
-			$clean['star_device_id']     = sanitize_text_field( $printer['star_device_id'] ?? '' );
-			$clean['star_client_type']   = sanitize_text_field( $printer['star_client_type'] ?? '' );
-		}
 
-		return $clean;
-	}
-
-	/**
-	 * Add server-owned client encoding fields for Star CloudPRNT printers.
-	 *
-	 * These fields let POS clients synthesize read-only cloud printer targets
-	 * without guessing how to render raw payloads before CloudPRNT delivery.
-	 *
-	 * @param array $printer Printer row.
-	 *
-	 * @return array
-	 */
-	private function with_cloud_printer_encoding_fields( array $printer ): array {
-		if ( 'star-cloudprnt' !== ( $printer['provider'] ?? '' ) ) {
-			return $printer;
-		}
-
-		$language = \in_array( $printer['language'] ?? '', array( 'esc-pos', 'star-prnt', 'star-line' ), true )
-			? $printer['language'] : 'esc-pos';
-		$columns  = isset( $printer['columns'] ) ? (int) $printer['columns'] : 42;
-		if ( ! \in_array( $columns, array( 32, 42, 48 ), true ) ) {
-			$columns = 42;
-		}
-
-		$printer['columns']           = $columns;
-		$printer['language']          = $language;
-		$printer['autoCut']           = array_key_exists( 'autoCut', $printer ) ? rest_sanitize_boolean( $printer['autoCut'] ) : true;
-		$printer['fullReceiptRaster'] = array_key_exists( 'fullReceiptRaster', $printer ) ? rest_sanitize_boolean( $printer['fullReceiptRaster'] ) : false;
-
-		return $printer;
+		return new WP_REST_Response( $result, 200 );
 	}
 
 	/**
 	 * Sanitize a cloud assignment entry.
 	 *
+	 * Kept for backward compatibility — the Settings_CloudPrint_Test conformance
+	 * gate exercises this method directly via ReflectionMethod. Delegates to
+	 * Cloud_Print_Section::sanitize_assignment() so the schema has exactly one
+	 * owner.
+	 *
 	 * @param mixed $assignment Assignment.
 	 *
 	 * @return array
+	 *
+	 * @phpstan-ignore-next-line
 	 */
 	private function sanitize_cloud_assignment( $assignment ): array {
-		$assignment = \is_array( $assignment ) ? $assignment : array();
+		$section = SettingsService::instance()->sections()->get( 'cloud_print' );
 
-		return array(
-			'printer_id'  => sanitize_text_field( $assignment['printer_id'] ?? '' ),
-			'store_id'    => isset( $assignment['store_id'] ) ? (int) $assignment['store_id'] : 0,
-			'scope'       => \in_array( $assignment['scope'] ?? '', array( 'every', 'pos', 'online' ), true ) ? $assignment['scope'] : 'every',
-			'template_id' => sanitize_text_field( (string) ( $assignment['template_id'] ?? '' ) ),
-		);
+		if ( $section instanceof \WCPOS\WooCommercePOS\Services\Settings\Cloud_Print_Section ) {
+			return $section->sanitize_assignment( $assignment );
+		}
+
+		return \is_array( $assignment ) ? $assignment : array();
 	}
